[es2017-conversion] In progress: Moved the THREE plugin files to the lib folder
