+++ src/js/lib/three-extras/VRControlsHyperbolic.js
