+++ src/js/lib/three-extras/VREffect.js
